Move Task model to ProjectAutomation
