+++ projects/tuist/fixtures/app_with_tasks/.gitignore
@@ -69,4 +69,4 @@ Derived/
 ### Tuist managed dependencies ###
 Tuist/Dependencies
 
-file.txt
+*.txt
